live_patch -> live_redirect for main tabs
My impression is that live_patch is intended only for links within the same LiveView (e.g. changing sort/filters) and not to other LiveViews.

When I experimented with it, they both seemed to behave the same, which I think comes down to live_patch turning into a live_redirect if it detects there's a new LiveView involved. Source for that idea: https://elixirforum.com/t/concrete-examples-of-when-to-use-live-patch-live-redirect-push-redirect-push-patch/30751/11

Presumably it's slightly less efficient, and expresses intent worse, to let live_patch fall back.