Persistent Settings
Doesn't seem like you can trivially store it in the session, because we only get passed session data on mount, and thereonafter we've only got a LiveView process via WebSocket – not a 'real' HTTP request/response cycle.

Using hooks like this is one way. Another could be to set a user identifier in the session and then store/retrieve data in the DB or Redis via that identifier. But that seemed more complicated and we'd want to consider expiring stale data etc.

There may be a 'stash' concept coming some day that would see something like this built into LiveView: https://news.ycombinator.com/item?id=21101081