Show reviewer Gravatar; store their email instead of Author
I don't understand why previous versions of this review tool would use Author to represent the reviewer. It's not an 'author', and it only appears we need the email.

It's slightly wasteful of data to store repeated emails, of course, but it seems to be the simplest thing that works for how we use it.