Get rid of Author and payload
Realised we just need a few fields. We've used some version of this app for 6+ years now and I don't think we've ever even discussed using more data than this. And if we do want it, we can always retrieve it later via API (or only store it for new data).