Unpolished beginnings of Comments and CommentNotifications
A lot to tidy up, but here's a first draft.

The commit_comment webhook only includes the username and integer ID.
So we're going to have to introduce a username field to the settings page.

Would be cool to some day use GitHub auth to get these details and
control access, but I doubt we'll bother.

Previous versions of this app stuck username, name and email in an
Author schema, but I think that would be more trouble than it's worth.
We had some bugs from it.