Commits: Show author usernames instead of committer name
So e.g. 'usera and userb' instead of 'Foo Ason and Bar Bson'.

The main reason is to be consistent with how we show names on the Comments tab. We don't get full names from the GitHub comment webhooks, so we'd need to make extra requests to show those on comments (or try to get them via commit records with a matching username).

Also, usernames are likely to be shorter, which is nice in a UI that's meant to be shown in a sidebar.