Broadcast when comments change state, so other clients update
Keeping this very simple. If it's plenty fast enough in real use this way (and I think it is), then life will be much easier if we can just reload from DB than if we have to keep enough state around and apply enough logic to remove a commit from your list if it suddenly falls outside your current filters, and so on.