Simplify connection detection: 1 socket, 0 channels
Seems like we can just rely on socket reconnection and not need to involve channel rejoining.

Renaming the socket for clarity.

Didn't seem like .phx-disconnected kicked in quickly on Heroku when turning off my Internet, so added a .consocket-closed.

Haven't looked into why, but at a guess, .phx-disconnected relies on the LiveSocket heartbeat which is something like 30 seconds.
When I tried in dev, turning off the server, it kicked in quickly, but perhaps then it's a clean close, where the
server has time to tell the client it's going.