Restore ping-based connection detection
Thought I could be clever and get away with something simpler, but it didn't work well in production.

Basically, if I turn off wifi with the Heroku app, it takes about 50 seconds until the (now offline) client considers the WebSocket to be dead, and then triggers the reconnection/rejoin we wanted to detect. That's too slow.

A lower client heartbeat won't affect this. The heartbeat is about sending stuff from client to server to keep the connection alive; the timeouts the browser/OS/network stack/whatever uses to determine that a WebSocket has died are independent of this, and probably not within my control.

Now that it doesn't have a special heartbeat, I'm going back to a (in future potentially) general purpose UserSocket.