Use URL params for comment filters
So a page reload, or switching tabs and back again, keeps the filters.

This has a bug: Since we re-mount on changed params, we'll reset the
username to the original username in the session, which could be stale.
But I plan on rewriting this all to be a single LiveView with
components, and then the username state will live in the top LiveView.
Should fix this issue.