Extract Commit.add_date_separators; overwrite previous separator
The old code didn't do this overwriting, so if commit X was previously the latest commit on day A, and then commit Y arrived via webhook, both X and Y could have had the day A separator on it.

Should also be faster (have not measured) by using a single map_reduce instead of chunk + flat map.