Don't hide date separator for today
Commits are shown sorted newest-ID-first rather than latest-commit-first.
This is because a commit from 10 days ago could be pushed today, and
should then appear at the top of the list.

But this also means that the first commit in the list could be from an
earlier date, and commits from today could come further down. In that
situation, we want the separator to show up. The easiest solution is to
always show it.