Introduce 'unlisted' commits
If a comment arrives by webhook but we don't have its commit in DB, the plan is to import the commit as 'unlisted'.

This lets us use the commit to display the comment, or to notify other commenters, without it suddenly appearing as a commit needing review.

We don't need a corresponding flag on comments. They're only 'listed' if we've created a CommentNotification anyway.