Fetch missing commits from GitHub when a comment arrives
Can easily happen if you comment on a commit that preceded the
introduction of Remit, or that has been removed from its DB since.

Before this change, that would cause a DB integrity error, because the
foreign key in the comment wouldn't have anything to point to.

This does not fetch any comments on the commit, so any previous
commenters remain un-notified. May fix that next.

Testing the API module with the HTTP lib's (Tesla's) built-in mocking.

Testing IngestComment, which calls the API module, with Mox mocking as
described in https://medium.com/flatiron-labs/elixir-test-mocking-with-mox-b825a955143f.

Refs #3.