Fetch comments along with missing commits
So those commenters are notified.

We use a transaction because if we create the commit but not its
comments, re-running this code wouldn't add those comments – they'd be
lost to us forever.

But we don't use a transaction for the ingest as a whole. If we e.g.
create the missing commit with its comments, but later fail to create
the ingested comment, that's fine – if we retry by e.g. re-sending the
webhook, it would skip the already-created commit-with comments but
would happily ingest the new comment.