Set a layout max width
The UI is intended to be shown in a narrow sidebar, and this
approximates that if you look at it in a regular browser.

I don't personally use it outside Fluid.app's sidebar in real use, but
if nothing else it may be good for development to have it look more
sidebar-like.

I picked a max-width that I think is bigger than anyone will actually
use in a sidebar, because otherwise it would annoyiny.