Attempt to speed up Comments.list_notifications query
I've noticed it feeling slow in the UI, and we got some exceptions related to slow queries and running out of pool: https://app.honeybadger.io/projects/72231/faults/65071008

I didn't do a proper EXPLAIN ANALYZE (maybe if this doesn't help) but from some experimentation, it seems resolved_at might be the main culprit. We filter on it (resolved/unresolved), but more significantly (I think) we sort on it.