Add index on comment_notifications.comment_id
Let's see if this improves the speed of Comments.list_notifications.

EXPLAIN ANALYZE doesn't seem to suggest the joining is slow, but I'm
grasping at straws and thinking that maybe somehow being able to join
with an index may help it actually use the LOWER(username) index – which
does seem to be the most expensive part.