Moar index
In an attempt to make Comments.list_notifications faster.

Asked in the #ecto Slack and someone pointed out that EXPLAIN pointed to a seq scan on LOWER(comments.commenter_username) (which wasn't indexed), so my index on LOWER(comment_notifications.username) wasn't likely to help :D

Let's see how we do with this one.