Decrease max_comments for performance
I think UI-wise it should be fine – 150 is plenty, and if you have more, you'll see the rest after you've resolved these ones.

Longer term it would be nice to have pagination behind the scenes, with an endless-scrolling UI.

This query was rather slow (100-150 ms) and I had no great luck with EXPLAIN ANALYZE and adding indexes. But experiments suggest that the query halves to ~50-75 ms with half the number of lines.