Show when you've reviewed too long
For now, we just check this whenever the list happens to update – e.g. when new commits come in or someone changes a commit's state.

Later, I may add a timer to check this even if no other activity happens.

Also snuck in some cleanup.

As with the link to the oldest unreviewed, it'd be nice if it actually opened that commit on click. See #24.

I made `overlong_in_review_by_me` return a list because some day we may want to link to all of them, not just the last.

Closes #37.