Keep unreviewed commits when periodically removing old data
Since we want to review all commits, and it seems a reasonable assumption for other projects as well.

Would be easy enough to also preserve commits with unresolved comment notifications, but it could look a bit weird to see a single old, reviewed commit in the commit list that's kept around just for its comments. So if we want to keep those forever, too, perhaps we should de-list the commit but keep it in DB.