lint: long lines
Also disabling this for tests because it's violated in a place where
fixing this makes it look _worse_, and I think long lines can be a more
natural occurrence in test code sometimes.

There is also an option to suppress specific checks locally if this is
deemed to be an overreach.