broadcast session changes internally
The main motivator for this was the Fluid user experience.

Fluid will open the login link in the right panel, and the auth flow
will complete there, leaving the left panel unaware that it is now
logged in until you reload it. This is clunky, especially if you've
hidden the nav controls.

This broadcasts the session changes to all running LiveView procs of the
same browser session. The session ID is generated early to make sure
there is always something identifiable there.

The logout changes are mainly there to keep both flows symmetric, but it
also makes the views reflect the real session state rather the view's
local copy of it.